requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
django-rest-swagger
gunicorn
gevent
whitenoise
django-filter